Cleans up agentUpdater a bit
